refactor(API error handling): separate error handling functions - add note on error string checking - remove error string checking for 401 as the error should be reserved for authentication only and no guarantee the string is consistent due to i18n
